typescript(vx-demo): rewrite Drag-ii demo in TypeScript
