+++ packages/vx-demo/src/components/gallery.js
@@ -36,7 +36,7 @@ import Treemap from './tiles/Treemap.tsx';
 import Radar from './tiles/radar';
 import Responsive from './tiles/Responsive.tsx';
 import DragI from './tiles/Drag-i.tsx';
-import DragII from './tiles/drag-ii';
+import DragII from './tiles/Drag-ii.tsx';
 import LinkTypes from './tiles/LinkTypes.tsx';
 import Threshold from './tiles/Threshold.tsx';
 import Chord from './tiles/Chord.tsx';
